requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bitarray==2.9.2
ccxt==4.3.52
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
configobj==5.0.8
cryptography==42.0.8
cytoolz==0.12.3
eth-abi==3.0.1
eth-account==0.8.0
eth-hash==0.7.0
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.1
hexbytes==0.3.1
hyperliquid-python-sdk==0.4.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.8
msgpack-python==0.5.6
multidict==6.0.5
numpy==1.26.4
orjson==3.10.5
pandas==2.2.2
parsimonious==0.8.1
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.32.3
rich==13.7.1
rlp==3.0.0
scipy==1.12.0
six==1.16.0
toolz==0.12.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
psutil==6.0.0
aiofiles==24.1.0
aiozmq==1.0.0
posix-ipc==1.1.1
pycurl==7.45.3
scikit-learn==1.4.2
lz4==4.3.3
python-binance==1.0.19
selenium==4.29.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
openai==1.43.0
aiohttp==3.10.10
httpcore==1.0.5
httpx==0.27.0
pymongo==4.11.2
